Created different calendar modes and added support for adapter interchangability without adapter state loss.
