+++ library/src/main/java/com/prolificinteractive/materialcalendarview/CalendarMode.java
@@ -1,0 +1,7 @@
+package com.prolificinteractive.materialcalendarview;
+
+public enum CalendarMode {
+
+    MONTHS,
+    WEEKS
+}
